@@ -1,20 +1,42 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Step, Stepper, StepLabel } from 'material-ui/Stepper';
-import RaisedButton from 'material-ui/RaisedButton';
-import FlatButton from 'material-ui/FlatButton';
-import getMuiTheme from 'material-ui/styles/getMuiTheme';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import DropDownMenu from 'material-ui/DropDownMenu';
-import MenuItem from 'material-ui/MenuItem';
-import Dialog from 'material-ui/Dialog';
+import {
+  Button, Step, Stepper, StepLabel, Select, MenuItem,
+  Dialog, DialogActions, DialogContent,
+} from '@material-ui/core';
+import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles';
 
+import ConfirmationModal from '../ConfirmationModal';
 import Search from '../Search';
 import api from '../../services/api';
 import util from '../../services/util';
 
-const muiTheme = getMuiTheme({
-  fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+const muiTheme = createMuiTheme({
+  palette: {
+    primary: {
+      main: '#0097a7',
+    },
+  },
+  typography: {
+    fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+  },
+  overrides: {
+    MuiStepper: {
+      root: {
+        padding: '24px 0px',
+      },
+    },
+    MuiButton: {
+      root: {
+        marginRight: '15px',
+      },
+    },
+    MuiFormControl: {
+      root: {
+        marginBottom: '15px',
+      },
+    },
+  },
 });
 
 const style = {
@@ -47,6 +69,8 @@ export default class AttachAddon extends Component {
       app: '',
       addons: [],
       addon: {},
+      loadingError: false,
+      loadingErrorMessage: '',
     };
   }
 
@@ -68,6 +92,8 @@ export default class AttachAddon extends Component {
           addon: response.data[0],
           loading: false,
         });
+      }).catch(() => {
+        this.setState(prevState, () => this.setState({ loadingErrorMessage: 'Could not find specified app', loadingError: true }));
       });
     }
   }
@@ -78,9 +104,9 @@ export default class AttachAddon extends Component {
         className={addon.addon_service.name}
         key={addon.name}
         value={addon}
-        label={addon.addon_service.name}
-        primaryText={addon.addon_service.name}
-      />
+      >
+        {addon.addon_service.name}
+      </MenuItem>
     ));
   }
 
@@ -89,7 +115,14 @@ export default class AttachAddon extends Component {
       case 0:
         return (
           <div>
-            <Search className={'app-search'} label="App" data={util.filterName(this.state.apps)} handleSearch={this.handleSearch} searchText="" errorText={this.state.errorText} />
+            <Search
+              className={'app-search'}
+              label="App"
+              data={util.filterName(this.state.apps)}
+              handleSearch={this.handleSearch}
+              errorText={this.state.errorText}
+              color="black"
+            />
             <p>
               The application name that has an addon you want to attach. Ex. my-test-app-dev
             </p>
@@ -98,9 +131,9 @@ export default class AttachAddon extends Component {
       case 1:
         return (
           <div>
-            <DropDownMenu className="addon-menu" value={this.state.addon} onChange={this.handleAddonChange}>
+            <Select className="addon-menu" value={this.state.addon} onChange={this.handleAddonChange}>
               {this.getAddons()}
-            </DropDownMenu>
+            </Select>
             <p>
               Select the addon you want to attach.
             </p>
@@ -122,9 +155,9 @@ export default class AttachAddon extends Component {
     });
   }
 
-  handleAddonChange = (event, index, value) => {
+  handleAddonChange = (event) => {
     this.setState({
-      addon: value,
+      addon: event.target.value,
     });
   }
 
@@ -168,7 +201,7 @@ export default class AttachAddon extends Component {
 
   renderContent() {
     const { finished, stepIndex } = this.state;
-    const contentStyle = { margin: '0 16px', overflow: 'hidden' };
+    const contentStyle = { margin: '0 32px' };
     if (finished) {
       this.submitAddonAttachment();
     } else {
@@ -176,20 +209,25 @@ export default class AttachAddon extends Component {
         <div style={contentStyle}>
           <div>{this.getStepContent(stepIndex)}</div>
           <div style={style.buttons.div}>
-            {stepIndex > 0 && (<FlatButton
-              className="back"
-              label="Back"
-              disabled={stepIndex === 0}
-              onTouchTap={this.handlePrev}
-              style={style.buttons.Back}
-            />)}
             {stepIndex > 0 && (
-              <RaisedButton
+              <Button
+                className="back"
+                disabled={stepIndex === 0}
+                onClick={this.handlePrev}
+                style={style.buttons.Back}
+              >
+                Back
+              </Button>
+            )}
+            {stepIndex > 0 && (
+              <Button
+                variant="contained"
                 className="next"
-                label={stepIndex === 1 ? 'Finish' : 'Next'}
-                primary
-                onTouchTap={this.handleNext}
-              />
+                color="primary"
+                onClick={this.handleNext}
+              >
+                {stepIndex === 1 ? 'Finish' : 'Next'}
+              </Button>
             )}
           </div>
         </div>
@@ -201,7 +239,7 @@ export default class AttachAddon extends Component {
   render() {
     const { loading, stepIndex, finished } = this.state;
     return (
-      <MuiThemeProvider muiTheme={muiTheme}>
+      <MuiThemeProvider theme={muiTheme}>
         <div style={style.stepper}>
           <Stepper activeStep={stepIndex}>
             <Step>
@@ -219,17 +257,43 @@ export default class AttachAddon extends Component {
           <Dialog
             className="attach-addon-error"
             open={this.state.submitFail}
-            modal
-            actions={
-              <FlatButton
-                className="ok"
-                label="Ok"
-                primary
-                onTouchTap={this.handleClose}
-              />}
           >
-            {this.state.submitMessage}
+            <DialogContent>
+              {this.state.submitMessage}
+            </DialogContent>
+            <DialogActions>
+              <Button
+                className="ok"
+                color="primary"
+                onClick={this.handleClose}
+              >
+                Ok
+              </Button>
+            </DialogActions>
           </Dialog>
+          <Dialog
+            className="load-app-error"
+            open={this.state.loadingError}
+          >
+            <DialogContent>
+              {this.state.loadingErrorMessage}
+            </DialogContent>
+            <DialogActions>
+              <Button
+                className="ok"
+                color="primary"
+                onClick={() => this.setState({ loadingError: false, loadingErrorMessage: '' })}
+              >
+                Ok
+              </Button>
+            </DialogActions>
+          </Dialog>
+          <ConfirmationModal
+            open={this.state.loadingError}
+            onOk={() => this.setState({ loadingError: false, loadingErrorMessage: '' })}
+            message={this.state.loadingErrorMessage}
+            title="Error"
+          />
         </div>
       </MuiThemeProvider>
     );
