@@ -1,18 +1,37 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Step, Stepper, StepLabel } from 'material-ui/Stepper';
-import RaisedButton from 'material-ui/RaisedButton';
-import FlatButton from 'material-ui/FlatButton';
-import getMuiTheme from 'material-ui/styles/getMuiTheme';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import DropDownMenu from 'material-ui/DropDownMenu';
-import MenuItem from 'material-ui/MenuItem';
-import Dialog from 'material-ui/Dialog';
-
+import {
+  Step, Stepper, StepLabel, FormControl,
+  Select, MenuItem, Dialog, Button, Input,
+  DialogTitle, DialogActions, DialogContent,
+} from '@material-ui/core';
+import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
 import api from '../../services/api';
 
-const muiTheme = getMuiTheme({
-  fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+const muiTheme = createMuiTheme({
+  palette: {
+    primary: { main: '#0097a7' },
+  },
+  typography: {
+    fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+  },
+  overrides: {
+    MuiStepper: {
+      root: {
+        padding: '24px 0px',
+      },
+    },
+    MuiButton: {
+      root: {
+        marginRight: '15px',
+      },
+    },
+    MuiFormControl: {
+      root: {
+        marginBottom: '15px',
+      },
+    },
+  },
 });
 
 const style = {
@@ -43,8 +62,10 @@ export default class NewAddon extends Component {
       submitMessage: '',
       services: [],
       service: {},
+      serviceid: '',
       plans: [],
       plan: {},
+      planid: '',
     };
   }
 
@@ -65,6 +86,7 @@ export default class NewAddon extends Component {
         this.setState({
           plans: response.data,
           plan: response.data[0],
+          planid: response.data[0].id,
           loading: false,
         });
       });
@@ -76,10 +98,10 @@ export default class NewAddon extends Component {
       <MenuItem
         className={service.human_name}
         key={service.id}
-        value={service}
-        label={service.human_name}
-        primaryText={service.human_name}
-      />
+        value={service.id}
+      >
+        {service.human_name}
+      </MenuItem>
     ));
   }
 
@@ -88,10 +110,10 @@ export default class NewAddon extends Component {
       <MenuItem
         className={plan.human_name}
         key={plan.name}
-        value={plan}
-        label={plan.name}
-        primaryText={plan.name}
-      />
+        value={plan.id}
+      >
+        {plan.name}
+      </MenuItem>
     ));
   }
 
@@ -100,9 +122,17 @@ export default class NewAddon extends Component {
       case 0:
         return (
           <div>
-            <DropDownMenu className="service-menu" value={this.state.service} onChange={this.handleServiceChange}>
-              {this.getServices()}
-            </DropDownMenu>
+            <FormControl className="service-form">
+              <Select
+                autoWidth
+                className="service-menu"
+                value={this.state.service.id}
+                onChange={this.handleServiceChange}
+                input={<Input name="service" id="service-helper" />}
+              >
+                {this.getServices()}
+              </Select>
+            </FormControl>
             <p>
               Select the akkeris addon you would like to attach to your app.
             </p>
@@ -111,10 +141,18 @@ export default class NewAddon extends Component {
       case 1:
         return (
           <div>
-            <DropDownMenu className="plan-menu" value={this.state.plan} onChange={this.handlePlanChange}>
-              {this.getPlans()}
-            </DropDownMenu>
-            <div className="plan-info">
+            <FormControl className="plan-form">
+              <Select
+                autoWidth
+                className="plan-menu"
+                value={this.state.planid}
+                onChange={this.handlePlanChange}
+                input={<Input name="plan" id="plan-helper" />}
+              >
+                {this.getPlans()}
+              </Select>
+            </FormControl>
+            <div className="plan-info" style={{ marginBottom: '15px' }}>
               {this.state.plan.price && this.state.plan.price.cents !== 0 && (
                 <span className="plan-price">
                   <b>{this.formatPrice(this.state.plan.price.cents)}/mo</b>
@@ -151,16 +189,16 @@ export default class NewAddon extends Component {
     });
   }
 
-  handleServiceChange = (event, index, value) => {
-    this.setState({
-      service: value,
-    });
+  handleServiceChange = (event) => {
+    const serviceid = event.target.value;
+    const service = this.state.services.find(a => a.id === serviceid);
+    this.setState({ service, serviceid });
   }
 
-  handlePlanChange = (event, index, value) => {
-    this.setState({
-      plan: value,
-    });
+  handlePlanChange = (event) => {
+    const planid = event.target.value;
+    const plan = this.state.plans.find(a => a.id === planid);
+    this.setState({ plan, planid });
   }
 
   handleNext = () => {
@@ -200,25 +238,9 @@ export default class NewAddon extends Component {
     });
   }
 
-  submitAddon = () => {
-    api.createAddon(this.props.app, this.state.plan.id).then(() => {
-      this.props.onComplete('Addon Created');
-    }).catch((error) => {
-      this.setState({
-        submitMessage: error.response.data,
-        submitFail: true,
-        finished: false,
-        stepIndex: 0,
-        loading: false,
-        plans: [],
-        plan: {},
-      });
-    });
-  }
-
   renderContent() {
     const { finished, stepIndex } = this.state;
-    const contentStyle = { margin: '0 16px', overflow: 'hidden' };
+    const contentStyle = { margin: '0 32px', overflow: 'hidden' };
     if (finished) {
       this.submitAddon();
     } else {
@@ -226,19 +248,23 @@ export default class NewAddon extends Component {
         <div style={contentStyle}>
           <div>{this.getStepContent(stepIndex)}</div>
           <div style={style.buttons.div}>
-            {stepIndex > 0 && (<FlatButton
-              className="back"
-              label="Back"
-              disabled={stepIndex === 0}
-              onTouchTap={this.handlePrev}
-              style={style.buttons.Back}
-            />)}
-            <RaisedButton
+            {stepIndex > 0 && (
+              <Button
+                className="back"
+                disabled={stepIndex === 0}
+                onClick={this.handlePrev}
+              >
+                Back
+              </Button>
+            )}
+            <Button
+              variant="contained"
               className="next"
-              label={stepIndex === 1 ? 'Finish' : 'Next'}
-              primary
-              onTouchTap={this.handleNext}
-            />
+              color="primary"
+              onClick={this.handleNext}
+            >
+              {stepIndex === 1 ? 'Finish' : 'Next'}
+            </Button>
           </div>
         </div>
       );
@@ -249,7 +275,7 @@ export default class NewAddon extends Component {
   render() {
     const { loading, stepIndex, finished } = this.state;
     return (
-      <MuiThemeProvider muiTheme={muiTheme}>
+      <MuiThemeProvider theme={muiTheme}>
         <div style={style.stepper}>
           <Stepper activeStep={stepIndex}>
             <Step>
@@ -267,16 +293,16 @@ export default class NewAddon extends Component {
           <Dialog
             className="new-addon-error"
             open={this.state.submitFail}
-            modal
-            actions={
-              <FlatButton
-                className="ok"
-                label="Ok"
-                primary
-                onTouchTap={this.handleClose}
-              />}
           >
-            {this.state.submitMessage}
+            <DialogTitle>Error</DialogTitle>
+            <DialogContent>{this.state.submitMessage}</DialogContent>
+            <DialogActions>
+              <Button
+                className="ok"
+                color="primary"
+                onClick={this.handleClose}
+              >OK</Button>
+            </DialogActions>
           </Dialog>
         </div>
       </MuiThemeProvider>
