@@ -1,18 +1,21 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import getMuiTheme from 'material-ui/styles/getMuiTheme';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import RefreshIndicator from 'material-ui/RefreshIndicator';
-import { Table, TableBody, TableRow, TableRowColumn, TableHeader, TableHeaderColumn } from 'material-ui/Table';
-import Dialog from 'material-ui/Dialog';
-import FlatButton from 'material-ui/FlatButton';
-
 import SHA256 from 'crypto-js/sha256';
+import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles';
+import {
+  CircularProgress, Table, TableHead, TableBody, TableRow, TableCell, Button,
+  Dialog, DialogContent, DialogActions, DialogTitle,
+} from '@material-ui/core';
 
 import api from '../../services/api';
 
-const muiTheme = getMuiTheme({
-  fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+const muiTheme = createMuiTheme({
+  palette: {
+    primary: { main: '#0097a7' },
+  },
+  typography: {
+    fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+  },
 });
 
 const style = {
@@ -51,18 +54,21 @@ const style = {
 };
 
 function sillyfunc(prefix, input) {
+  if (!input) {
+    return null;
+  }
   return Object.keys(input).map((key) => {
     if (typeof input[key] === 'object') {
       return sillyfunc(prefix ? `${prefix}.${key}` : key, input[key]);
     }
     return (
-      <TableRow>
-        <TableRowColumn>
-          {prefix ? `${prefix}.${key}` : key}
-        </TableRowColumn>
-        <TableRowColumn>
-          {input[key]}
-        </TableRowColumn>
+      <TableRow key={key}>
+        <TableCell>
+          <span>{prefix ? `${prefix}.${key}` : key}</span>
+        </TableCell>
+        <TableCell>
+          <span>{input[key]}</span>
+        </TableCell>
       </TableRow>
     );
   });
@@ -91,25 +97,24 @@ export default class Audits extends Component {
   getAudits() {
     return this.state.audits.map((audit) => {
       const id = SHA256(JSON.stringify(audit)).toString().substring(0, 7);
-
       return (
         <TableRow
           className={id}
           key={id}
           style={style.tableRow}
-          onTouchTap={() => this.handleRowSelection(id)}
-          selectable={false}
+          onClick={() => this.handleRowSelection(id)}
+          hover
         >
-          <TableRowColumn style={style.tableRow}>
+          <TableCell style={style.tableRow}>
             <div style={style.tableRowColumn.main}>{audit.action}</div>
             <div style={style.tableRowColumn.sub}>{id}</div>
-          </TableRowColumn>
-          <TableRowColumn>
+          </TableCell>
+          <TableCell>
             <div>{audit.username}</div>
-          </TableRowColumn>
-          <TableRowColumn>
+          </TableCell>
+          <TableCell>
             <div>{new Date(audit.received_at).toLocaleString()}</div>
-          </TableRowColumn>
+          </TableCell>
         </TableRow>
       );
     });
@@ -122,6 +127,7 @@ export default class Audits extends Component {
         if (id === this.state.id) {
           return sillyfunc('', JSON.parse(audit.info));
         }
+        return null;
       })
     );
   }
@@ -136,53 +142,64 @@ export default class Audits extends Component {
   handleDialogClose = () => {
     this.setState({
       diagOpen: false,
-      id: '',
+      // id: '',
     });
   }
 
   render() {
     if (this.state.loading) {
       return (
-        <MuiThemeProvider muiTheme={muiTheme}>
+        <MuiThemeProvider theme={muiTheme}>
           <div style={style.refresh.div}>
-            <RefreshIndicator top={0} size={40} left={0} style={style.refresh.indicator} status="loading" />
+            <CircularProgress top={0} size={40} left={0} style={style.refresh.indicator} status="loading" />
           </div>
         </MuiThemeProvider>
       );
     }
     return (
-      <MuiThemeProvider muiTheme={muiTheme}>
+      <MuiThemeProvider theme={muiTheme}>
         <div>
-          <Table className="audit-list" wrapperStyle={{ overflow: 'visible' }} bodyStyle={{ overflow: 'visible' }}>
-            <TableHeader adjustForCheckbox={false} displaySelectAll={false} selectable={false}>
+          <Table className="audit-list" >
+            <TableHead>
               <TableRow>
-                <TableHeaderColumn>Change</TableHeaderColumn>
-                <TableHeaderColumn>User</TableHeaderColumn>
-                <TableHeaderColumn>Time</TableHeaderColumn>
+                <TableCell>
+                  <span>Change</span>
+                </TableCell>
+                <TableCell>
+                  <span>User</span>
+                </TableCell>
+                <TableCell >
+                  <span>Time</span>
+                </TableCell>
               </TableRow>
-            </TableHeader>
-            <TableBody displayRowCheckbox={false} showRowHover selectable={false}>
+            </TableHead>
+            <TableBody>
               {this.getAudits()}
             </TableBody>
           </Table>
           <Dialog
             className="audit-dialog"
             open={this.state.diagOpen}
-            autoScrollBodyContent
-            title="Audit Info"
-            actions={
-              <FlatButton
-                className="ok"
-                label="Ok"
-                primary
-                onTouchTap={this.handleDialogClose}
-              />}
+            // Clear id on exited to avoid race between
+            // closing dialog and clearing displayed audit info
+            onExited={() => { this.setState({ id: '' }); }}
+            maxWidth="lg"
           >
-            <Table className="audit-info">
-              <TableBody displayRowCheckbox={false} showRowHover={false} selectable={false}>
-                {this.getAuditInfo()}
-              </TableBody>
-            </Table>
+            <DialogTitle>Audit Info</DialogTitle>
+            <DialogContent>
+              <Table className="audit-info">
+                <TableBody>
+                  {this.getAuditInfo()}
+                </TableBody>
+              </Table>
+            </DialogContent>
+            <DialogActions>
+              <Button
+                className="ok"
+                color="primary"
+                onClick={this.handleDialogClose}
+              >Ok</Button>
+            </DialogActions>
           </Dialog>
         </div>
       </MuiThemeProvider>
