@@ -1,19 +1,22 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import getMuiTheme from 'material-ui/styles/getMuiTheme';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import RefreshIndicator from 'material-ui/RefreshIndicator';
-import { List, ListItem } from 'material-ui/List';
-import RaisedButton from 'material-ui/RaisedButton';
-import Dialog from 'material-ui/Dialog';
-import FlatButton from 'material-ui/FlatButton';
-import RemoveIcon from 'material-ui/svg-icons/content/clear';
+import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles';
+import {
+  Dialog, DialogContent, DialogTitle, DialogContentText, DialogActions,
+  List, ListItem, ListItemText, Button, CircularProgress,
+} from '@material-ui/core';
+import RemoveIcon from '@material-ui/icons/Clear';
 
 import api from '../../services/api';
 import ConfirmationModal from '../ConfirmationModal';
 
-const muiTheme = getMuiTheme({
-  fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+const muiTheme = createMuiTheme({
+  palette: {
+    primary: { main: '#0097a7' },
+  },
+  typography: {
+    fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+  },
 });
 
 const style = {
@@ -23,7 +26,8 @@ const style = {
       marginRight: 'auto',
       width: '40px',
       height: '40px',
-      marginTop: '20%',
+      marginTop: '15%',
+      marginBottom: '5%',
     },
     indicator: {
       display: 'inline-block',
@@ -32,6 +36,12 @@ const style = {
   },
   button: {
     marginRight: '16px',
+  },
+  deleteButtonLabel: {
+    paddingRight: '5px',
+  },
+  removeIcon: {
+    paddingRight: '5px',
   },
 };
 
@@ -90,37 +100,58 @@ export default class SiteOverview extends Component {
   render() {
     if (this.state.loading) {
       return (
-        <MuiThemeProvider muiTheme={muiTheme}>
+        <MuiThemeProvider theme={muiTheme}>
           <div style={style.refresh.div}>
-            <RefreshIndicator top={0} size={40} left={0} style={style.refresh.indicator} status="loading" />
+            <CircularProgress top={0} size={40} left={0} style={style.refresh.indicator} status="loading" />
           </div>
         </MuiThemeProvider>
       );
     }
     return (
-      <MuiThemeProvider muiTheme={muiTheme}>
+      <MuiThemeProvider theme={muiTheme}>
         <div>
           <List>
-            <ListItem primaryText="Domain" secondaryText={this.props.site.domain} disabled />
-            <ListItem primaryText="ID" secondaryText={this.props.site.id} disabled />
-            <ListItem primaryText="Region" secondaryText={this.props.site.region.name} disabled />
-            <ListItem primaryText="Compliance" secondaryText={this.props.site.compliance.toString()} disabled />
-            <ListItem primaryText="Updated At" secondaryText={Date(this.props.site.updated_at).toLocaleString()} rightIconButton={<RaisedButton className="delete" style={style.button}label="Delete Site" onTouchTap={this.handleConfirmation} secondary icon={<RemoveIcon />} />} disabled />
+            <ListItem >
+              <ListItemText primary="Domain" secondary={this.props.site.domain} />
+            </ListItem>
+            <ListItem >
+              <ListItemText primary="ID" secondary={this.props.site.id} />
+            </ListItem>
+            <ListItem >
+              <ListItemText primary="Region" secondary={this.props.site.region.name} />
+            </ListItem>
+            <ListItem >
+              <ListItemText primary="Compliance" secondary={this.props.site.compliance.toString()} />
+            </ListItem>
+            <ListItem >
+              <ListItemText primary="Updated At" secondary={Date(this.props.site.updated_at).toLocaleString()} />
+              <Button
+                variant="contained"
+                className="delete"
+                style={style.button}
+                onClick={this.handleConfirmation}
+                color="secondary"
+              >
+                <RemoveIcon nativeColor="white" style={style.removeIcon} />
+                <span style={style.deleteButtonLabel}>Delete Site</span>
+              </Button>
+            </ListItem>
           </List>
           <ConfirmationModal open={this.state.open} onOk={this.handleRemoveSite} onCancel={this.handleCancelConfirmation} message="Are you sure you want to delete this site?" />
           <Dialog
             className="error"
             open={this.state.submitFail}
-            modal
-            actions={
-              <FlatButton
-                className="ok"
-                label="Ok"
-                primary
-                onTouchTap={this.handleClose}
-              />}
           >
-            {this.state.submitMessage}
+            <DialogTitle>Error</DialogTitle>
+            <DialogContent>
+              <DialogContentText>{this.state.submitMessage}</DialogContentText>
+            </DialogContent>
+            <DialogActions>
+              <Button
+                color="primary"
+                onClick={this.handleClose}
+              >Ok</Button>
+            </DialogActions>
           </Dialog>
         </div>
       </MuiThemeProvider>
