@@ -1,13 +1,16 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Table, TableBody, TableRow, TableRowColumn, TableHeader, TableHeaderColumn } from 'material-ui/Table';
+import {
+  Table, TableBody, TableHead, TableRow, TableCell,
+} from '@material-ui/core';
+
 
 const style = {
   tableRow: {
     height: '58px',
     cursor: 'pointer',
   },
-  tableRowColumn: {
+  tableCell: {
     title: {
       fontSize: '16px',
     },
@@ -23,38 +26,44 @@ export default class SitesList extends Component {
     return this.props.sites.map((site) => {
       const date = new Date(site.updated_at);
       return (
-        <TableRow className={site.domain} key={site.id} style={style.tableRow}>
-          <TableRowColumn>
-            <div style={style.tableRowColumn.title}>{site.domain}</div>
-            <div style={style.tableRowColumn.sub}>{site.id}</div>
-          </TableRowColumn>
-          <TableRowColumn>
+        <TableRow
+          className={site.domain}
+          key={site.id}
+          style={style.tableRow}
+          hover
+          onClick={() => this.handleRowSelection(site.id)}
+        >
+          <TableCell>
+            <div style={style.tableCell.title}>{site.domain}</div>
+            <div style={style.tableCell.sub}>{site.id}</div>
+          </TableCell>
+          <TableCell>
             <div>{date.toLocaleString()}</div>
-          </TableRowColumn>
-          <TableRowColumn>
-            <div style={style.tableRowColumn.title}>{site.region.name}</div>
-          </TableRowColumn>
+          </TableCell>
+          <TableCell>
+            <div style={style.tableCell.title}>{site.region.name}</div>
+          </TableCell>
         </TableRow>
       );
     });
   }
 
-  handleRowSelection = (selectedRows) => {
-    window.location = `#/sites/${this.props.sites[selectedRows].domain}/info`;
+  handleRowSelection = (id) => {
+    window.location = `#/sites/${id}/info`;
   }
 
   render() {
     return (
       <div>
-        <Table className="site-list" onRowSelection={this.handleRowSelection} wrapperStyle={{ overflow: 'visible' }} bodyStyle={{ overflow: 'visible' }}>
-          <TableHeader adjustForCheckbox={false} displaySelectAll={false} selectable={false}>
+        <Table className="site-list">
+          <TableHead>
             <TableRow>
-              <TableHeaderColumn>Site</TableHeaderColumn>
-              <TableHeaderColumn>Updated</TableHeaderColumn>
-              <TableHeaderColumn>Region</TableHeaderColumn>
+              <TableCell>Site</TableCell>
+              <TableCell>Updated</TableCell>
+              <TableCell>Region</TableCell>
             </TableRow>
-          </TableHeader>
-          <TableBody displayRowCheckbox={false} showRowHover>
+          </TableHead>
+          <TableBody>
             {this.getSites()}
           </TableBody>
         </Table>
