@@ -1,23 +1,28 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Step, Stepper, StepLabel } from 'material-ui/Stepper';
-import RaisedButton from 'material-ui/RaisedButton';
-import FlatButton from 'material-ui/FlatButton';
-import getMuiTheme from 'material-ui/styles/getMuiTheme';
-import HelpIcon from 'material-ui/svg-icons/action/help';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import lightBaseTheme from 'material-ui/styles/baseThemes/lightBaseTheme';
-import TextField from 'material-ui/TextField';
-import Dialog from 'material-ui/Dialog';
-import Checkbox from 'material-ui/Checkbox';
-import { IconButton } from 'material-ui';
-import { GridList } from 'material-ui/GridList';
-
+import {
+  Step, Stepper, StepLabel, Button, Checkbox, Grid, TextField, IconButton, Typography,
+  Dialog, DialogTitle, DialogContent, DialogContentText, DialogActions, Tooltip, FormControlLabel,
+} from '@material-ui/core';
+import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles';
+import HelpIcon from '@material-ui/icons/Help';
 import api from '../../services/api';
 import eventDescriptions from './EventDescriptions.js'; // eslint-disable-line import/extensions
 
-const muiTheme = getMuiTheme({
-  fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+const muiTheme = createMuiTheme({
+  palette: {
+    primary: { main: '#0097a7' },
+  },
+  typography: {
+    fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+  },
+  overrides: {
+    MuiCheckbox: {
+      root: {
+        padding: '2px 12px',
+      },
+    },
+  },
 });
 
 const style = {
@@ -29,17 +34,16 @@ const style = {
     width: '70px',
   },
   eventsLabel: {
-    color: 'rgba(0, 0, 0, 0.3)',
-    fontSize: '12px',
+    // color: 'rgba(0, 0, 0, 0.3)',
+    fontSize: '14px',
+    paddingRight: '8px',
   },
-  checkAllActive: {
-    width: '25%',
+  checkAllContainer: {
     borderTop: '1px solid black',
-    marginTop: '10px',
-    paddingTop: '10px',
-    display: 'inline-block',
+    marginTop: '5px',
+    paddingTop: '5px',
   },
-  gridListWidth: {
+  gridContainer: {
     width: '350px',
   },
   stepper: {
@@ -61,17 +65,13 @@ const style = {
     paddingTop: '20px',
   },
   eventsInfoButton: {
-    icon: {
-      height: '18px', width: '18px',
-    },
     padding: '0',
-    height: '24px',
-    width: '24px',
+    height: '18px',
+    width: '18px',
   },
   eventsInfoIcon: {
     height: '18px',
     width: '18px',
-    color: lightBaseTheme.palette.accent1Color,
   },
 };
 
@@ -101,11 +101,12 @@ export default class NewWebhook extends Component {
           <div>
             <TextField
               className="webhook-url"
-              floatingLabelText="URL"
+              label="URL"
               type="text"
               value={this.state.url}
-              onChange={this.handleURLChange}
-              errorText={this.state.errorText}
+              onChange={this.handleChange('url')}
+              helperText={this.state.errorText ? this.state.errorText : ''}
+              error={!!this.state.errorText}
             />
             <p>
               Enter a URL for the new webhook (defaults to http).
@@ -117,32 +118,35 @@ export default class NewWebhook extends Component {
           <div>
             <div style={style.eventsHeader}>
               <p style={style.eventsLabel}>Events</p>
-              <IconButton
-                className="events-info-button"
-                onTouchTap={this.openEventsInfoDialog}
-                style={style.eventsInfoButton}
-                iconStyle={style.eventsInfoIcon}
-                tooltip="Click for Descriptions"
-                tooltipPosition="top-right"
-              >
-                <HelpIcon />
-              </IconButton>
+              <Tooltip placement="right" title="Click for Descriptions">
+                <IconButton
+                  className="events-info-button"
+                  onClick={this.openEventsInfoDialog}
+                  style={style.eventsInfoButton}
+                >
+                  <HelpIcon style={style.eventsInfoIcon} color="secondary" />
+                </IconButton>
+              </Tooltip>
             </div>
             {this.renderEventsInfoDialog()}
             <div className="events">
-              <GridList cellHeight="auto" style={style.gridListWidth}>
+              <Grid container spacing={8} style={style.gridContainer}>
                 {this.getEventCheckboxes(this.webhook)}
-              </GridList>
-              <span style={style.checkAllActive} >
-                <Checkbox
-                  label="Check All"
-                  value="Check All"
-                  key="Check All"
-                  className="checkbox-check-all"
-                  checked={this.state.checkedAll}
-                  onCheck={this.handleCheckAll}
-                />
-              </span>
+                <Grid item xs={12} style={style.checkAllContainer}>
+                  <FormControlLabel
+                    control={
+                      <Checkbox
+                        value="Check All"
+                        key="Check All"
+                        className="checkbox-check-all"
+                        checked={this.state.checkedAll}
+                        onChange={this.handleCheckAll}
+                      />
+                    }
+                    label="Check All"
+                  />
+                </Grid>
+              </Grid>
             </div>
             {this.state.errorText && (
               <div style={style.eventsError} className="events-errorText">
@@ -156,10 +160,11 @@ export default class NewWebhook extends Component {
           <div>
             <TextField
               className="webhook-secret"
-              floatingLabelText="Secret"
+              label="Secret"
               type="password"
               value={this.state.secret}
-              onChange={this.handleSecretChange}
+              onChange={this.handleChange('secret')}
+              helperText={this.state.errorText ? this.state.errorText : ''}
               errorText={this.state.errorText}
             />
             <p>
@@ -167,6 +172,9 @@ export default class NewWebhook extends Component {
             </p>
           </div>
         );
+        // Have to have this otherwise it displays "Error- Captain Hook not found" on submit
+      case 3:
+        return '';
       default:
         return 'Error- Captain Hook not found';
     }
@@ -174,15 +182,20 @@ export default class NewWebhook extends Component {
 
   getEventCheckboxes() { // eslint-disable-line class-methods-use-this
     return defaultEvents.map(event => (
-      <Checkbox
-        className={`checkbox-${event}`}
-        key={event}
-        value={event}
-        label={event}
-        checked={this.state.events.includes(event)}
-        onCheck={this.handleCheck}
-        style={style.checkboxWidth}
-      />
+      <Grid key={`checkbox-${event}`} item xs={6}>
+        <FormControlLabel
+          control={
+            <Checkbox
+              className={`checkbox-${event}`}
+              key={event}
+              value={event}
+              checked={this.state.events.includes(event)}
+              onChange={this.handleCheck}
+            />
+          }
+          label={event}
+        />
+      </Grid>
     ));
   }
 
@@ -223,16 +236,8 @@ export default class NewWebhook extends Component {
     this.setState({ events: currEvents });
   }
 
-  handleURLChange = (event, value) => {
-    this.setState({
-      url: value,
-    });
-  }
-
-  handleSecretChange = (event, value) => {
-    this.setState({
-      secret: value,
-    });
+  handleChange = name => (event) => {
+    this.setState({ [name]: event.target.value });
   }
 
   handleClose = () => {
@@ -298,32 +303,29 @@ export default class NewWebhook extends Component {
 
   renderEventsInfoDialog() {
     return (
-      <Dialog
-        className="events-info-dialog"
-        open={this.state.eventsDialogOpen}
-        title="Description of Events"
-        autoScrollBodyContent
-        actions={
-          <FlatButton
+      <Dialog open={this.state.eventsDialogOpen} className="events-info-dialog">
+        <DialogTitle>Description of Events</DialogTitle>
+        <DialogContent>
+          <div>
+            {eventDescriptions.data.map((event, index) => (
+              <p key={`${event}.length`}><b>{defaultEvents[index]}</b><br />{event}</p>
+            ))}
+          </div>
+        </DialogContent>
+        <DialogActions>
+          <Button
             className="ok"
-            label="Ok"
-            primary
-            onTouchTap={this.closeEventsInfoDialog}
-          />
-        }
-      >
-        <div>
-          {eventDescriptions.data.map((event, index) => (
-            <p key={`${event}.length`}><b>{defaultEvents[index]}</b><br />{event}</p>
-          ))}
-        </div>
+            color="primary"
+            onClick={this.closeEventsInfoDialog}
+          >Ok</Button>
+        </DialogActions>
       </Dialog>
     );
   }
 
   renderContent() {
     const { finished, stepIndex } = this.state;
-    const contentStyle = { margin: '0 16px', overflow: 'visible' };
+    const contentStyle = { margin: '0 57px', overflow: 'visible' };
     if (finished) {
       this.submitWebHook();
     } else {
@@ -331,19 +333,20 @@ export default class NewWebhook extends Component {
         <div style={contentStyle}>
           <div>{this.getStepContent(stepIndex)}</div>
           <div style={style.buttons.div}>
-            {stepIndex > 0 && (<FlatButton
-              className="back"
-              label="Back"
-              disabled={stepIndex === 0}
-              onTouchTap={this.handlePrev}
-              style={style.buttons.back}
-            />)}
-            <RaisedButton
+            {stepIndex > 0 && (
+              <Button
+                className="back"
+                disabled={stepIndex === 0}
+                onClick={this.handlePrev}
+                style={style.buttons.back}
+              >Back</Button>
+            )}
+            <Button
+              variant="contained"
               className="next"
-              label={stepIndex === 2 ? 'Finish' : 'Next'}
-              primary
-              onTouchTap={this.handleNext}
-            />
+              color="primary"
+              onClick={this.handleNext}
+            >{stepIndex === 2 ? 'Finish' : 'Next'}</Button>
           </div>
         </div>
       );
@@ -354,7 +357,7 @@ export default class NewWebhook extends Component {
   render() {
     const { stepIndex } = this.state;
     return (
-      <MuiThemeProvider muiTheme={muiTheme}>
+      <MuiThemeProvider theme={muiTheme}>
         <div style={style.stepper}>
           <Stepper activeStep={stepIndex}>
             <Step>
@@ -370,25 +373,27 @@ export default class NewWebhook extends Component {
           <div>
             {this.renderContent()}
           </div>
-          <Dialog
-            className="new-webhook-error"
-            open={this.state.submitFail}
-            modal
-            actions={
-              <FlatButton
-                className="ok"
-                label="Ok"
-                primary
-                onTouchTap={this.handleClose}
-              />}
-          >
-            {this.state.submitMessage}
+          <Dialog open={this.state.submitFail} className="new-webhook-error">
+            <DialogTitle>
+              <Typography variant="h6">
+                Error
+              </Typography>
+            </DialogTitle>
+            <DialogContent>
+              <DialogContentText>
+                {this.state.submitMessage}
+              </DialogContentText>
+            </DialogContent>
+            <DialogActions>
+              <Button label="Ok" color="primary" onClick={this.handleClose}>Ok</Button>
+            </DialogActions>
           </Dialog>
         </div>
       </MuiThemeProvider>
     );
   }
 }
+
 
 NewWebhook.propTypes = {
   app: PropTypes.string.isRequired,
