@@ -1,14 +1,11 @@
 import React, { Component } from 'react';
-import { Toolbar, ToolbarGroup } from 'material-ui/Toolbar';
+import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles';
+import {
+  Toolbar, IconButton, Select, MenuItem,
+  CircularProgress, Paper, FormControl, InputLabel,
+} from '@material-ui/core';
 import { Link } from 'react-router-dom';
-import IconButton from 'material-ui/IconButton';
-import getMuiTheme from 'material-ui/styles/getMuiTheme';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import DropDownMenu from 'material-ui/DropDownMenu';
-import MenuItem from 'material-ui/MenuItem';
-import RefreshIndicator from 'material-ui/RefreshIndicator';
-import Paper from 'material-ui/Paper';
-import AddIcon from 'material-ui/svg-icons/content/add';
+import AddIcon from '@material-ui/icons/Add';
 
 import api from '../../services/api';
 import AppList from '../../components/Apps/AppList';
@@ -17,25 +14,72 @@ import Search from '../../components/Search';
 
 /* eslint-disable jsx-a11y/anchor-is-valid */
 
-const muiTheme = getMuiTheme({
-  fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+const muiTheme = createMuiTheme({
+  palette: {
+    primary: { main: '#0097a7' },
+  },
+  typography: {
+    fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+  },
+  overrides: {
+    MuiIconButton: {
+      root: { color: 'white', padding: '6px', marginBottom: '-6px' },
+    },
+    MuiToolbar: {
+      root: {
+        minHeight: '48px !important',
+        maxHeight: '48px !important',
+      },
+    },
+    MuiInputLabel: {
+      root: { color: 'white !important' },
+      shrink: { color: 'white !important' },
+      animated: { color: 'white !important' },
+    },
+    MuiSelect: {
+      root: { color: 'white' },
+      icon: { color: 'white' },
+      select: { color: 'white !important' },
+    },
+    MuiInput: {
+      input: {
+        '&::placeholder': {
+          color: 'white',
+        },
+        color: 'white',
+      },
+      underline: {
+        // Border color when input is not selected
+        '&:before': {
+          borderBottom: '1px solid rgb(200, 200, 200)',
+        },
+        // Border color when input is selected
+        '&:after': {
+          borderBottom: '2px solid white',
+        },
+        // Border color on hover
+        '&:hover:not([class^=".MuiInput-disabled-"]):not([class^=".MuiInput-focused-"]):not([class^=".MuiInput-error-"]):before': {
+          borderBottom: '1px solid rgb(200, 200, 200)',
+        },
+      },
+    },
+  },
 });
 
 const style = {
-  filter: {
-    paddingLeft: '0px',
-  },
   refresh: {
     div: {
       marginLeft: 'auto',
       marginRight: 'auto',
       width: '40px',
       height: '40px',
-      marginTop: '20%',
+      marginTop: '15%',
+      marginBottom: '5%',
     },
     indicator: {
       display: 'inline-block',
       position: 'relative',
+      color: 'white',
     },
   },
   toolbar: {
@@ -44,12 +88,13 @@ const style = {
     marginLeft: 'auto',
     marginRight: 'auto',
     padding: '16px 0',
+    display: 'flex',
+    flexDirection: 'row',
+    alignItems: 'flex-end',
   },
   link: {
     textDecoration: 'none',
-  },
-  font: {
-    color: 'white',
+    marginLeft: 'auto',
   },
   paper: {
     maxWidth: '1024px',
@@ -57,20 +102,13 @@ const style = {
     marginRight: 'auto',
     marginTop: '12px',
   },
-  menu: {
-    height: '48px',
-    lineHeight: '48px',
-    color: 'white',
+  spaceContainer: {
+    marginLeft: '30px',
+    minWidth: '145px',
   },
-  icon: {
-    top: '0px',
-  },
-  search: {
-    color: 'white',
-    WebkitTextFillColor: 'white',
-  },
-  searchHint: {
-    color: 'rgba(255,255,255,0.3)',
+  regionContainer: {
+    marginLeft: '30px',
+    minWidth: '145px',
   },
 };
 
@@ -78,8 +116,8 @@ export default class Apps extends Component {
   constructor(props, context) {
     super(props, context);
     this.state = {
-      space: 'all',
-      region: 'all',
+      space: '',
+      region: '',
       apps: [],
       filteredApps: [],
       filteredSpaces: [],
@@ -116,13 +154,13 @@ export default class Apps extends Component {
 
   getSpaces() {
     return this.state.filteredSpaces.map(space => (
-      <MenuItem className={space.name} key={space.id} value={space.name} label={`Space: ${space.name}`} primaryText={space.name} />
+      <MenuItem className={space.name} key={space.id} value={space.name}>{space.name}</MenuItem>
     ));
   }
 
   getRegions() {
     return this.state.regions.map(region => (
-      <MenuItem className={region.name} key={region.id} value={region.name} label={`Region: ${region.name}`} primaryText={region.name} />
+      <MenuItem className={region.name} key={region.id} value={region.name}>{region.name}</MenuItem>
     ));
   }
 
@@ -130,8 +168,8 @@ export default class Apps extends Component {
     window.location = `#/apps/${searchText}/info`;
   }
 
-  handleSpaceChange = (event, index, value) => {
-    const space = value;
+  handleSpaceChange = (event) => {
+    const space = event.target.value;
     const apps = util.filterApps(this.state.apps, space);
     this.setState({
       space,
@@ -139,14 +177,14 @@ export default class Apps extends Component {
     });
   }
 
-  handleRegionChange = (event, index, value) => {
-    const region = value;
+  handleRegionChange = (event) => {
+    const region = event.target.value;
     const apps = util.filterAppsByRegion(this.state.apps, region);
     const spaces = util.filterSpacesByRegion(this.state.spaces, region);
 
     this.setState({
-      space: 'all',
-      region: value,
+      space: '',
+      region: event.target.value,
       filteredApps: apps,
       filteredSpaces: spaces,
     });
@@ -155,52 +193,54 @@ export default class Apps extends Component {
   render() {
     if (this.state.loading) {
       return (
-        <MuiThemeProvider muiTheme={muiTheme}>
+        <MuiThemeProvider theme={muiTheme}>
           <div className="loading" style={style.refresh.div}>
-            <RefreshIndicator top={0} size={40} left={0} style={style.refresh.indicator} status="loading" />
+            <CircularProgress top={0} size={40} left={0} style={style.refresh.indicator} status="loading" />
           </div>
         </MuiThemeProvider>);
     }
     return (
-      <MuiThemeProvider muiTheme={muiTheme}>
+      <MuiThemeProvider theme={muiTheme}>
         <div>
-          <Toolbar style={style.toolbar}>
-            <ToolbarGroup>
-              <Search
-                style={style.search}
-                hintStyle={style.searchHint}
-                data={util.filterName(this.state.filteredApps)}
-                handleSearch={this.handleSearch}
-                className="search"
-              />
-              <DropDownMenu
+          <Toolbar style={style.toolbar} disableGutters>
+            <Search
+              data={util.filterName(this.state.filteredApps)}
+              handleSearch={this.handleSearch}
+              className="search"
+            />
+            <FormControl style={style.regionContainer}>
+              <InputLabel htmlFor="region-select">Filter by Region</InputLabel>
+              <Select
                 className="region-dropdown"
-                style={style.filter}
-                labelStyle={style.menu}
-                iconStyle={style.icon}
                 value={this.state.region}
                 onChange={this.handleRegionChange}
+                inputProps={{
+                  name: 'region',
+                  id: 'region-select',
+                }}
               >
-                <MenuItem className="all" value="all" label="Filter by Region" primaryText="all" />
+                <MenuItem className="all" value="all">All</MenuItem>
                 {this.getRegions()}
-              </DropDownMenu>
-              <DropDownMenu
+              </Select>
+            </FormControl>
+            <FormControl style={style.spaceContainer}>
+              <InputLabel htmlFor="space-select">Filter by Space</InputLabel>
+              <Select
                 className="space-dropdown"
-                style={style.filter}
-                labelStyle={style.menu}
-                iconStyle={style.icon}
                 value={this.state.space}
                 onChange={this.handleSpaceChange}
+                inputProps={{
+                  name: 'space',
+                  id: 'space-select',
+                }}
               >
-                <MenuItem className="all" value="all" label="Filter by Space" primaryText="all" />
+                <MenuItem className="all" value="all">All</MenuItem>
                 {this.getSpaces()}
-              </DropDownMenu>
-            </ToolbarGroup>
-            <ToolbarGroup>
-              <Link to="/apps/new" style={style.link}>
-                <IconButton className="new-app" iconStyle={style.font}><AddIcon /></IconButton>
-              </Link>
-            </ToolbarGroup>
+              </Select>
+            </FormControl>
+            <Link to="/apps/new" style={style.link}>
+              <IconButton className="new-app"><AddIcon /></IconButton>
+            </Link>
           </Toolbar>
           <Paper style={style.paper}>
             <AppList className="apps" apps={this.state.filteredApps} />
