@@ -1,17 +1,19 @@
 import React, { Component } from 'react';
-import getMuiTheme from 'material-ui/styles/getMuiTheme';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
+import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles';
+import {
+  Paper, List, ListSubheader, CircularProgress, LinearProgress,
+} from '@material-ui/core';
 import PropTypes from 'prop-types';
-import RefreshIndicator from 'material-ui/RefreshIndicator';
-import LinearProgress from 'material-ui/LinearProgress';
-import { List } from 'material-ui/List';
-import Paper from 'material-ui/Paper';
-import Subheader from 'material-ui/Subheader';
 import api from '../../services/api';
 import recommendations from '../../services/util/recommendations';
 
-const muiTheme = getMuiTheme({
-  fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+const muiTheme = createMuiTheme({
+  palette: {
+    primary: { main: '#0097a7' },
+  },
+  typography: {
+    fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+  },
 });
 
 /* eslint-disable no-console */
@@ -23,11 +25,13 @@ const style = {
       marginRight: 'auto',
       width: '40px',
       height: '40px',
-      marginTop: '20%',
+      marginTop: '15%',
+      marginBottom: '5%',
     },
     indicator: {
       display: 'inline-block',
       position: 'relative',
+      color: 'white',
     },
   },
   toolbar: {
@@ -151,8 +155,9 @@ export default class WasteReport extends Component {
     });
   }
 
-  getApps() {
-    return this.state.apps.map((app) => {
+  getAppRecommendations() {
+    const results = [];
+    this.state.apps.forEach((app) => {
       const notes = recommendations.execute(
         app.metrics,
         app.formations,
@@ -160,35 +165,44 @@ export default class WasteReport extends Component {
         this.state.sizes,
       );
       if (notes.length > 0) {
-        return (
+        results.push(
           <div key={`recommend_${app.name}`} className="recommendations" style={style.recommendations}>
-            <Subheader>{app.name}</Subheader>
+            <ListSubheader>{app.name}</ListSubheader>
             <ul>
               {notes}
             </ul>
-          </div>
+          </div>,
         );
       }
-      return null;
     });
+    if (results.length === 0) {
+      results.push(
+        <div key={'reccomend_empty'} style={style.recommendations}>
+          <ListSubheader>
+            No recommendations available for {this.props.match.params.org}
+          </ListSubheader>
+        </div>,
+      );
+    }
+    return results;
   }
 
   render() {
     if (this.state.loading) {
       return (
-        <MuiThemeProvider muiTheme={muiTheme}>
+        <MuiThemeProvider theme={muiTheme}>
           <div className="invoices" style={style.refresh.div}>
-            <RefreshIndicator top={0} size={40} left={0} style={style.refresh.indicator} status="loading" />
+            <CircularProgress top={0} size={40} left={0} style={style.refresh.indicator} status="loading" />
           </div>
         </MuiThemeProvider>);
     } else if (this.state.analyzing) {
       return (
-        <MuiThemeProvider muiTheme={muiTheme}>
+        <MuiThemeProvider theme={muiTheme}>
           <div className="invoices">
             <Paper style={style.paper}>
               <div className="internal">
                 <span style={style.label}>Analyzing...</span>
-                <LinearProgress mode="determinate" value={this.state.progress} />
+                <LinearProgress variant="determinate" value={this.state.progress} />
               </div>
             </Paper>
           </div>
@@ -196,12 +210,12 @@ export default class WasteReport extends Component {
       );
     }
     return (
-      <MuiThemeProvider muiTheme={muiTheme}>
+      <MuiThemeProvider theme={muiTheme}>
         <div className="invoices">
           <Paper style={style.paper}>
             <div style={style.header}>Cost Report for {this.props.match.params.org}</div>
             <List>
-              {this.getApps()}
+              {this.getAppRecommendations()}
             </List>
           </Paper>
         </div>
